+++ projects/ngx-tools/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @consensus-labs/ngx-tools
 
+## 1.0.14
+
+### Patch Changes
+
+- Remove overflow from harmonica animation
+
 ## 1.0.13
 
 ### Patch Changes